# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai:complete:
        post:
            tags:
                - AIService
            description: |-
                Complete runs one writing-assistant turn against an OpenAI-compatible
                 chat-completions endpoint configured on the instance.
            operationId: AIService_Complete
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CompleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CompleteResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:synthesize:
        post:
            tags:
                - AIService
            description: Synthesize converts text to speech audio using an instance AI provider.
            operationId: AIService_Synthesize
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SynthesizeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SynthesizeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        CompleteRequest:
            required:
                - prompt
            type: object
            properties:
                prompt:
                    type: string
                    description: Required. The user instruction for this turn (e.g. "polish this paragraph").
                content:
                    type: string
                    description: |-
                        Optional. Source text to transform — usually the editor selection, or the
                         whole draft when no selection exists.
        CompleteResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The model's reply text. Writing-assistant responses are plain markdown.
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        SynthesizeRequest:
            required:
                - text
            type: object
            properties:
                text:
                    type: string
                    description: Required. Text to convert to speech.
        SynthesizeResponse:
            type: object
            properties:
                audio:
                    type: string
                    description: Generated audio bytes.
                    format: bytes
                contentType:
                    type: string
                    description: MIME type of the generated audio, e.g. "audio/mpeg".
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
        TranscribeResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The transcribed text.
        TranscriptionAudio:
            type: object
            properties:
                content:
                    writeOnly: true
                    type: string
                    description: Inline audio bytes.
                    format: bytes
                uri:
                    type: string
                    description: URI for audio content. Reserved for future use.
                filename:
                    type: string
                    description: Optional. The uploaded filename.
                contentType:
                    type: string
                    description: Optional. The MIME type of the input audio.
...